Inject ConnectionFactory in Jdbc based connectors
