+++ presto-postgresql/src/main/java/io/prestosql/plugin/postgresql/PostgreSqlClientModule.java
@@ -15,9 +15,14 @@ package io.prestosql.plugin.postgresql;
 
 import com.google.inject.Binder;
 import com.google.inject.Module;
+import com.google.inject.Provides;
 import com.google.inject.Scopes;
+import com.google.inject.Singleton;
 import io.prestosql.plugin.jdbc.BaseJdbcConfig;
+import io.prestosql.plugin.jdbc.ConnectionFactory;
+import io.prestosql.plugin.jdbc.DriverConnectionFactory;
 import io.prestosql.plugin.jdbc.JdbcClient;
+import org.postgresql.Driver;
 
 import static io.airlift.configuration.ConfigBinder.configBinder;
 
@@ -30,5 +35,12 @@ public class PostgreSqlClientModule
         binder.bind(JdbcClient.class).to(PostgreSqlClient.class).in(Scopes.SINGLETON);
         configBinder(binder).bindConfig(BaseJdbcConfig.class);
         configBinder(binder).bindConfig(PostgreSqlConfig.class);
+    }
+
+    @Provides
+    @Singleton
+    public ConnectionFactory getConnectionFactory(BaseJdbcConfig config)
+    {
+        return new DriverConnectionFactory(new Driver(), config);
     }
 }
