+++ presto-redshift/src/main/java/io/prestosql/plugin/redshift/RedshiftClientModule.java
@@ -15,9 +15,14 @@ package io.prestosql.plugin.redshift;
 
 import com.google.inject.Binder;
 import com.google.inject.Module;
+import com.google.inject.Provides;
 import com.google.inject.Scopes;
+import com.google.inject.Singleton;
 import io.prestosql.plugin.jdbc.BaseJdbcConfig;
+import io.prestosql.plugin.jdbc.ConnectionFactory;
+import io.prestosql.plugin.jdbc.DriverConnectionFactory;
 import io.prestosql.plugin.jdbc.JdbcClient;
+import org.postgresql.Driver;
 
 import static io.airlift.configuration.ConfigBinder.configBinder;
 
@@ -29,5 +34,12 @@ public class RedshiftClientModule
     {
         binder.bind(JdbcClient.class).to(RedshiftClient.class).in(Scopes.SINGLETON);
         configBinder(binder).bindConfig(BaseJdbcConfig.class);
+    }
+
+    @Singleton
+    @Provides
+    public ConnectionFactory getConnectionFactory(BaseJdbcConfig config)
+    {
+        return new DriverConnectionFactory(new Driver(), config);
     }
 }
